♻️MusicPlayerViewController 리팩토링
- final 키워드 추가
- private 추가
- private을 추가하기 위해 CollectionView+MusicPlayerViewController 파일 삭제
