+++ JipJung/JipJung/UI/Sound/Views/MusicPlayerViewController.swift
@@ -7,36 +7,31 @@
 
 import AVKit
 import UIKit
-
 import SnapKit
+import RxSwift
+import RxCocoa
 
-class MusicPlayerViewController: UIViewController {
-    // MARK: - View builder -> 임시 코드가 포함됨
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .lightContent
-    }
+final class MusicPlayerViewController: UIViewController {
+    private let dataSource = ["Relax", "Melody", "Meditation", "Etc"]
+    private let themeColor = CGColor(red: 34.0/255.0, green: 48.0/255.0, blue: 74.0/255.0, alpha: 1)
     
-    let dataSource = ["Relax", "Melody", "Meditation", "Etc"]
-    let themeColor = CGColor(red: 34.0/255.0, green: 48.0/255.0, blue: 74.0/255.0, alpha: 1)
+    private let navigationBar = UINavigationBar()
+    private let backButton = BackCircleButton()
+    private let favoriteButton = FavoriteCircleButton()
     
-    let navigationBar = UINavigationBar()
-    let backButton = BackCircleButton()
-    let favoriteButton = FavoriteCircleButton()
-    
-    let topView = UIView()
-    let musicDescriptionView = MusicDescriptionView()
-    let tagCollectionView: UICollectionView = {
+    private let topView = UIView()
+    private let musicDescriptionView = MusicDescriptionView()
+    private let tagCollectionView: UICollectionView = {
         let layout = LeftAlignCollectionViewFlowLayout()
-        
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
         collectionView.backgroundColor = .clear
         collectionView.register(TagCollectionViewCell.self, forCellWithReuseIdentifier: UICollectionView.CellIdentifier.tag.value)
         return collectionView
     }()
-    
-    let bottomView = UIView()
-    let maximTextView = MusicPlayerMaximView()
-    let playButton: UIButton = {
+
+    private let bottomView = UIView()
+    private let maximTextView = MusicPlayerMaximView()
+    private let playButton: UIButton = {
         let button = UIButton(type: .system)
         button.titleLabel?.font = .boldSystemFont(ofSize: 16)
         button.backgroundColor = .white
@@ -44,25 +39,30 @@ class MusicPlayerViewController: UIViewController {
         button.layer.cornerRadius = 8
         return button
     }()
-
-    var audioPlayer: AVAudioPlayer?
     
-    var playerLooper: AVPlayerLooper?
-    var queuePlayer: AVQueuePlayer?
-    var playerLayer: AVPlayerLayer?
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return .lightContent
+    }
+
+    // MARK: - Private Variables
+    
+    private var viewModel: MusicPlayerViewModel?
+    private var disposeBag: DisposeBag = DisposeBag()
+    private var audioPlayer: AVAudioPlayer?
+    private var playerLooper: AVPlayerLooper?
+    private var queuePlayer: AVQueuePlayer?
+    private var playerLayer: AVPlayerLayer?
     
     // MARK: - Life Cycles
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         self.view.backgroundColor = .white
 
         configureUI()
-        
+        bindUI()
         tagCollectionView.delegate = self
         tagCollectionView.dataSource = self
-        playButton.addTarget(self, action: #selector(playButtonTouched(_:)), for: .touchUpInside)
-        backButton.addTarget(self, action: #selector(backButtonTouched(_:)), for: .touchUpInside)
-        favoriteButton.addTarget(self, action: #selector(favoriteButtonTouched(_:)), for: .touchUpInside)
     }
     
     override func viewDidLayoutSubviews() {
@@ -70,33 +70,20 @@ class MusicPlayerViewController: UIViewController {
         playerLayer?.frame = topView.bounds
     }
     
-    func configureUI() {
+    // MARK: - Initializer
+
+    convenience init(viewModel: MusicPlayerViewModel) {
+        self.init(nibName: nil, bundle: nil)
+        self.viewModel = viewModel
+    }
+    
+    private func configureUI() {
         configureTopView()
         configureBottomView()
         configureNavigationBar()
     }
     
-    // MARK: - Input vent from views
-    @objc func favoriteButtonTouched(_ sender: UIButton) {
-        print(#function, #line)
-    }
-    
-    @objc func streamingButtonTouched(_ sender: UIButton) {
-        print(#function, #line)
-    }
-    
-    @objc func playButtonTouched(_ sender: UIButton) {
-        print(#function, #line)
-    }
-    
-    @objc func backButtonTouched(_ sender: UIButton) {
-        navigationController?.popViewController(animated: true)
-    }
-}
-
-// MARK: - UI Configuration Method
-extension MusicPlayerViewController {
-    func configureNavigationBar() {
+    private func configureNavigationBar() {
         self.view.addSubview(navigationBar)
         navigationBar.snp.makeConstraints { make in
             make.leading.top.trailing.equalTo(view.safeAreaLayoutGuide)
@@ -120,7 +107,7 @@ extension MusicPlayerViewController {
         }
     }
     
-    func configureTopView() {
+    private func configureTopView() {
         configureVideoView()
                 
         self.view.addSubview(topView)
@@ -142,7 +129,7 @@ extension MusicPlayerViewController {
         }
     }
     
-    func configureVideoView() {
+    private func configureVideoView() {
         guard let videoURL = Bundle.main.url(forResource: "sea", withExtension: "mp4")
         else {
             return
@@ -172,7 +159,7 @@ extension MusicPlayerViewController {
         queuePlayer.play()
     }
     
-    func configureBottomView() {
+    private func configureBottomView() {
         bottomView.layer.backgroundColor = themeColor
         playButton.setTitleColor(UIColor(cgColor: themeColor), for: .normal)
         
@@ -205,4 +192,62 @@ extension MusicPlayerViewController {
         }
     }
     
+    private func bindUI() {
+    }
+}
+
+extension MusicPlayerViewController: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return dataSource.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(
+            withReuseIdentifier: UICollectionView.CellIdentifier.tag.value,
+            for: indexPath
+        ) as? TagCollectionViewCell
+        else {
+            return UICollectionViewCell()
+        }
+        cell.tagLabel.text = dataSource[indexPath.item]
+        return cell
+    }
+}
+
+extension MusicPlayerViewController: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let label = UILabel()
+        label.font = .systemFont(ofSize: 14)
+        label.text = dataSource[indexPath.item]
+        label.sizeToFit()
+        
+        let size = label.frame.size
+        
+        return CGSize(width: size.width + 16, height: size.height + 10)
+    }
+}
+
+class LeftAlignCollectionViewFlowLayout: UICollectionViewFlowLayout {
+    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
+        let attributes = super.layoutAttributesForElements(in: rect)
+        
+        var leftMargin = sectionInset.left
+        var maxY: CGFloat = -1.0
+        
+        attributes?.forEach { layoutAttribute in
+          if layoutAttribute.representedElementCategory == .cell {
+            if layoutAttribute.frame.origin.y >= maxY {
+              leftMargin = sectionInset.left
+            }
+
+            layoutAttribute.frame.origin.x = leftMargin
+            leftMargin += layoutAttribute.frame.width + minimumInteritemSpacing
+            maxY = max(layoutAttribute.frame.maxY, maxY)
+          }
+        }
+        
+        return attributes
+      }
 }
